# attached_assets/Gita-Reader/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gita/chapters/{chapterId}:
    get:
      operationId: getChapter
      tags: [content]
      summary: Get a chapter with verse list
      description: Returns full chapter detail including list of verse summaries
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 18
      responses:
        "200":
          description: Chapter detail with verses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterDetailResponse"
        "404":
          description: Chapter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gita/chapters/{chapterId}/verses/{verseId}:
    get:
      operationId: getVerse
      tags: [content]
      summary: Get a single verse with full detail
      description: Returns complete verse information including Sanskrit, transliteration, Hindi/English meanings, word-by-word breakdown, and detailed explanation. AI-generated and cached if not pre-built.
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 18
        - name: verseId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Full verse detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerseDetailResponse"
        "404":
          description: Verse not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /gita/verses/explain-line:
    post:
      operationId: explainVerseLine
      tags: [gita]
      summary: Explain a specific line of a verse
      description: Get a detailed word-by-word and contextual explanation of a specific line from a Gita verse in Hindi and English
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExplainLineBody"
      responses:
        "200":
          description: Detailed line explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LineExplanation"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    VerseSummary:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        skt:
          type: string
          description: First line of Sanskrit verse
        iast:
          type: string
          description: Roman transliteration (first line)
        english:
          type: string
          description: Short English translation
        theme:
          type: string
      required:
        - id
        - chapterId
        - skt
        - iast
        - english
        - theme
    ChapterDetailResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        skt:
          type: string
        meaning:
          type: string
        totalVerses:
          type: integer
        summary:
          type: string
        longSummary:
          type: string
        themes:
          type: array
          items:
            type: string
        keyVerseRef:
          type: string
        setting:
          type: string
        verses:
          type: array
          items:
            $ref: "#/components/schemas/VerseSummary"
      required:
        - id
        - name
        - skt
        - meaning
        - totalVerses
        - summary
        - longSummary
        - themes
        - keyVerseRef
        - setting
        - verses
    WordMeaning:
      type: object
      properties:
        word:
          type: string
          description: Sanskrit word in Devanagari
        iast:
          type: string
          description: Roman transliteration
        hindi:
          type: string
          description: Hindi meaning
        english:
          type: string
          description: English meaning
      required:
        - word
        - iast
        - hindi
        - english
    VerseDetailResponse:
      type: object
      properties:
        id:
          type: integer
          description: Verse number within the chapter
        chapterId:
          type: integer
        chapterName:
          type: string
        skt:
          type: string
          description: Full Sanskrit verse in Devanagari
        iast:
          type: string
          description: Full Roman transliteration
        hindi:
          type: string
          description: Hindi translation
        english:
          type: string
          description: English translation
        wordByWord:
          type: array
          items:
            $ref: "#/components/schemas/WordMeaning"
        explanation:
          type: string
          description: Detailed explanation (Gita Press style) in Hindi
        gitaPressNote:
          type: string
          description: Gita Press specific commentary note
        modernRelevance:
          type: string
          description: Modern day relevance and application
        themes:
          type: array
          items:
            type: string
        prevVerse:
          type: integer
          nullable: true
          description: Previous verse number, null if first
        nextVerse:
          type: integer
          nullable: true
          description: Next verse number, null if last
        isAiGenerated:
          type: boolean
          description: Whether this detail was AI-generated (vs pre-built)
      required:
        - id
        - chapterId
        - chapterName
        - skt
        - iast
        - hindi
        - english
        - wordByWord
        - explanation
        - gitaPressNote
        - modernRelevance
        - themes
        - isAiGenerated
    ExplainLineBody:
      type: object
      properties:
        chapterId:
          type: integer
        verseId:
          type: integer
        line:
          type: string
          description: The specific Sanskrit line to explain
        context:
          type: string
          description: Optional full verse context
      required:
        - chapterId
        - verseId
        - line
    LineExplanation:
      type: object
      properties:
        line:
          type: string
        hindi:
          type: string
          description: Hindi explanation of this line
        english:
          type: string
          description: English explanation of this line
        wordBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/WordMeaning"
          description: Word-by-word breakdown of this specific line
      required:
        - line
        - hindi
        - english
        - wordBreakdown