# attached_assets/Gita-Reader/lib/api-zod/src/generated/types/chapterDetailResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VerseSummary } from "./verseSummary";

export interface ChapterDetailResponse {
  id: number;
  name: string;
  skt: string;
  meaning: string;
  totalVerses: number;
  summary: string;
  longSummary: string;
  themes: string[];
  keyVerseRef: string;
  setting: string;
  verses: VerseSummary[];
}