# attached_assets/Gita-Reader/lib/api-zod/src/generated/types/explainLineBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExplainLineBody {
  chapterId: number;
  verseId: number;
  /** The specific Sanskrit line to explain */
  line: string;
  /** Optional full verse context */
  context?: string;
}