# attached_assets/Gita-Reader/lib/api-zod/src/generated/types/lineExplanation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WordMeaning } from "./wordMeaning";

export interface LineExplanation {
  line: string;
  /** Hindi explanation of this line */
  hindi: string;
  /** English explanation of this line */
  english: string;
  /** Word-by-word breakdown of this specific line */
  wordBreakdown: WordMeaning[];
}