# attached_assets/Gita-Reader/lib/api-zod/src/generated/types/verseDetailResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WordMeaning } from "./wordMeaning";

export interface VerseDetailResponse {
  /** Verse number within the chapter */
  id: number;
  chapterId: number;
  chapterName: string;
  /** Full Sanskrit verse in Devanagari */
  skt: string;
  /** Full Roman transliteration */
  iast: string;
  /** Hindi translation */
  hindi: string;
  /** English translation */
  english: string;
  wordByWord: WordMeaning[];
  /** Detailed explanation (Gita Press style) in Hindi */
  explanation: string;
  /** Gita Press specific commentary note */
  gitaPressNote: string;
  /** Modern day relevance and application */
  modernRelevance: string;
  themes: string[];
  /** Previous verse number, null if first */
  prevVerse?: number | null;
  /** Next verse number, null if last */
  nextVerse?: number | null;
  /** Whether this detail was AI-generated (vs pre-built) */
  isAiGenerated: boolean;
}