# attached_assets/Gita-Reader/lib/api-zod/src/generated/types/verseSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VerseSummary {
  id: number;
  chapterId: number;
  /** First line of Sanskrit verse */
  skt: string;
  /** Roman transliteration (first line) */
  iast: string;
  /** Short English translation */
  english: string;
  theme: string;
}