# attached_assets/Gita-Reader/lib/api-zod/src/generated/types/wordMeaning.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WordMeaning {
  /** Sanskrit word in Devanagari */
  word: string;
  /** Roman transliteration */
  iast: string;
  /** Hindi meaning */
  hindi: string;
  /** English meaning */
  english: string;
}